requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.0
pydantic-settings==2.1.0
jinja2==3.1.2
aiofiles==23.2.1
weasyprint==60.2
pillow==10.1.0
razorpay==1.4.1
python-dotenv==1.0.0
email-validator==2.1.0
itsdangerous==2.1.2
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1